Adjustment of classnames
Adjusted classnames of checked radio buttons to singlechoice specific nomenclature in order to prevent interference with other plugins or standard moodle classnames.

Also implemented a javascript line that not only toggles the prop attribute to check/uncheck a radio but manually adds and removes the "checked" attribute to/from a input element inside the dom.